First pass at a smart toolbar. (it largely doesn't work) The toolbar tries to be smart, it sits at the top of the paragraph and only displays: 1. The buttons that are currently selected 2. The buttons that are most frequently used
When you roll over the toolbar it displays more options, but tries to keep the location of your current options the same for easy clicks.

When you create a blank section (new line) the toolbar displays additional "Card" options as well as block level changes, so you can insert an image, insert a markdown snippet, or html embed, as well as creating a list and a heading, etc

This is not styled and is no where near the final version.